[WP-13] Modify build name and refactor folders, files and firebase connection
