+++ app/build.gradle
@@ -7,7 +7,7 @@ android {
     compileSdk 31
 
     defaultConfig {
-        applicationId "com.example.wapps"
+        applicationId "com.catrenat.wapps"
         minSdk 21
         targetSdk 31
         versionCode 1
