+++ app/src/androidTest/java/com/catrenat/wapps/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package com.example.wapps;
+package com.catrenat.wapps;
 
 import android.content.Context;
