+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.example.wapps">
+    package="com.catrenat.wapps">
 
     <uses-permission android:name="android.permission.VIBRATE" />
     <application
@@ -11,7 +11,7 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.Wapps">
         <activity
-            android:name=".LoginScreen.LoginScreen"
+            android:name="com.catrenat.wapps.LoginScreen.LoginScreen"
             android:windowSoftInputMode="adjustPan"
             android:exported="true">
             <intent-filter>
@@ -21,10 +21,10 @@
             </intent-filter>
         </activity>
         <activity
-            android:name=".MainActivity"
+            android:name="com.catrenat.wapps.MainActivity"
             android:exported="false"/>
         <activity
-            android:name=".RegisterActivity"
+            android:name="com.catrenat.wapps.RegisterActivity"
             android:exported="false"/>
     </application>
