+++ app/src/main/java/com/catrenat/wapps/MainActivity.java
@@ -1,4 +1,4 @@
-package com.example.wapps;
+package com.catrenat.wapps;
 
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AppCompatActivity;
