+++ app/src/main/java/com/catrenat/wapps/MusicFragment.java
@@ -1,4 +1,4 @@
-package com.example.wapps;
+package com.catrenat.wapps;
 
 import android.os.Bundle;
 
@@ -8,12 +8,12 @@ import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 
-public class BooksFragment extends Fragment {
+public class MusicFragment extends Fragment {
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_books, container, false);
+        return inflater.inflate(R.layout.fragment_music, container, false);
     }
 }
