+++ app/src/main/res/layout/activity_main.xml
@@ -10,7 +10,7 @@
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/fragment_container"
-        android:name="com.example.wapps.GeneralFragment"
+        android:name="com.catrenat.wapps.GeneralFragment"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         app:layout_constraintBottom_toTopOf="@+id/bottom_navigation"
